refactor storage tests
